delete solver.py and merge to intervals.py
